#62 make child classes of LightingEngine to extract specific update algorithms
