+++ src/main/java/com/lordstephen77/dreamblocks/light/LightingPoint.java
@@ -1,4 +1,6 @@
-package com.lordstephen77.dreamblocks;
+package com.lordstephen77.dreamblocks.light;
+
+import com.lordstephen77.dreamblocks.Direction;
 
 import java.util.Comparator;
